QcomModulePkg: Add androidboot.*params to bootconfig
In Android S (AOSP experimental), the boot-config
feature replaces the androidboot.* kernel cmd-line
options in use with Android 11 and lower. The
boot-config feature is a mechanism for passing
configuration details from the build and bootloader
to Android S (AOSP experimental).  This feature provides
a way to separate the configuration parameters for
Android user-space from those for the kernel. Moving
the lengthy androidboot.* kernel parameters to the
boot-config file creates space on the kernel cmd-line
and makes it available for easy future expansion.

Change-Id: Ia0d5e7f2a7cec01b028597932ee95dc3dfa2cbe2