QcomModulePkg: Add functionality for appended recovery images
With GKI 2.0, we are compiling recovery as a separate ramdisk to be
appended in between the boot and vendor_boot images. This adds the
changes necessary for ensuring that image gets appended and verified
correctly, as well as for ensuring that builds which have a recovery
ramdisk with a kernel are loaded when necessary.

Change-Id: I457002e9681d122f5d57922bfa3366021520c368
Signed-off-by: Melody Olvera <molvera@codeaurora.org>